Original_With_Reps
On the Codecademy Loops Lesson, I was required to create my own for/else statement from scratch and so I decided to have some fun with it. I ended up creating a very simple 2016 South Carolina Primary simulator from just the Democratic side because including both parties turned out to be more difficult based on the logic I was using. Sourcing some historical turnout percentages and then creating if/else statements on the basis of "high" turnout vs "low" turnout, I was able to make it so the program would return Sanders as the winner when turnout was high and Clinton when turnout was low.

Sidenote: I will probably try to come back later and add comments to explain my logic throughout the code.