Update Original_With_Reps
Left Republican voter turnout and percentages while taking out the section where a Republican winner would be picked. The reasoning behind this is that it was too difficult for me to create code to pick a Rep winner using the same logic I did with the Dems. There were too many Rep candidates.